// providers/socket-types.tsx
export interface AttachmentsSendingProps {
	file: File;
	_id: string;
	name: string;
	preview: string;
	size: string | number;
	type: string;
	progress?: number;
	createdAt?: string;
}

export interface AttachmentsResponseProps {
	_id: string;
	name?: string;
	type?: string;
	size?: string | number;
	url?: string;
	uploadProgress?: number;
}

// Create a type that makes all properties optional
export type CombinedAttachmentType = Partial<AttachmentsResponseProps> & Partial<AttachmentsSendingProps>;